requirements.txt
numpy<2
openai-whisper
ollama
click